Move test utility code to test directory.
