+++ container-search/src/main/java/com/yahoo/search/dispatch/searchcluster/SearchGroupsImpl.java
@@ -2,10 +2,7 @@ package com.yahoo.search.dispatch.searchcluster;
 
 import com.google.common.math.Quantiles;
 
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -39,26 +36,5 @@ public class SearchGroupsImpl implements SearchGroups {
         if (isEmpty()) return 0;
         var activeDocuments = groups().stream().map(Group::activeDocuments).collect(Collectors.toList());
         return (long) Quantiles.median().compute(activeDocuments);
-    }
-
-
-    public static SearchGroupsImpl buildGroupListForTest(int numGroups, int nodesPerGroup, double minActivedocsPercentage) {
-        return new SearchGroupsImpl(buildGroupMapForTest(numGroups, nodesPerGroup), minActivedocsPercentage);
-    }
-    public static Map<Integer, Group> buildGroupMapForTest(int numGroups, int nodesPerGroup) {
-        Map<Integer, Group> groups = new HashMap<>();
-        int distributionKey = 0;
-        for (int group = 0; group < numGroups; group++) {
-            List<Node> groupNodes = new ArrayList<>();
-            for (int i = 0; i < nodesPerGroup; i++) {
-                Node node = new Node(distributionKey, "host" + distributionKey, group);
-                node.setWorking(true);
-                groupNodes.add(node);
-                distributionKey++;
-            }
-            Group g = new Group(group, groupNodes);
-            groups.put(group, g);
-        }
-        return Map.copyOf(groups);
     }
 }
