+++ container-search/src/test/java/com/yahoo/search/dispatch/SearchPathTest.java
@@ -2,8 +2,8 @@
 package com.yahoo.search.dispatch;
 
 import com.yahoo.search.dispatch.SearchPath.InvalidSearchPathException;
+import com.yahoo.search.dispatch.searchcluster.MockSearchCluster;
 import com.yahoo.search.dispatch.searchcluster.SearchGroups;
-import com.yahoo.search.dispatch.searchcluster.SearchGroupsImpl;
 import com.yahoo.search.dispatch.searchcluster.Node;
 import org.junit.jupiter.api.Test;
 
@@ -92,7 +92,7 @@ public class SearchPathTest {
 
     @Test
     void searchPathMustFilterNodesBasedOnDefinition() {
-        SearchGroups cluster = SearchGroupsImpl.buildGroupListForTest(3, 3, 100);
+        SearchGroups cluster = MockSearchCluster.buildGroupListForTest(3, 3, 100);
 
         assertEquals(distKeysAsString(SearchPath.selectNodes("1/1", cluster)), "4");
         assertEquals(distKeysAsString(SearchPath.selectNodes("/1", cluster)), "3,4,5");
