+++ container-search/src/test/java/com/yahoo/search/dispatch/searchcluster/MockSearchCluster.java
@@ -1,10 +1,12 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.search.dispatch;
+package com.yahoo.search.dispatch.searchcluster;
 
-import com.yahoo.search.dispatch.searchcluster.SearchGroupsImpl;
-import com.yahoo.search.dispatch.searchcluster.Node;
-import com.yahoo.search.dispatch.searchcluster.SearchCluster;
 import com.yahoo.vespa.config.search.DispatchConfig;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 /**
  * @author ollivir
@@ -12,7 +14,7 @@ import com.yahoo.vespa.config.search.DispatchConfig;
 public class MockSearchCluster extends SearchCluster {
 
     public MockSearchCluster(String clusterId, int groups, int nodesPerGroup) {
-        super(clusterId, SearchGroupsImpl.buildGroupListForTest(groups, nodesPerGroup, 88.0), null, null);
+        super(clusterId, buildGroupListForTest(groups, nodesPerGroup, 88.0), null, null);
     }
 
     @Override
@@ -48,6 +50,26 @@ public class MockSearchCluster extends SearchCluster {
             builder.maxWaitAfterCoverageFactor(0.5);
         }
         return builder;
+    }
+
+    public static SearchGroupsImpl buildGroupListForTest(int numGroups, int nodesPerGroup, double minActivedocsPercentage) {
+        return new SearchGroupsImpl(buildGroupMapForTest(numGroups, nodesPerGroup), minActivedocsPercentage);
+    }
+    private static Map<Integer, Group> buildGroupMapForTest(int numGroups, int nodesPerGroup) {
+        Map<Integer, Group> groups = new HashMap<>();
+        int distributionKey = 0;
+        for (int group = 0; group < numGroups; group++) {
+            List<Node> groupNodes = new ArrayList<>();
+            for (int i = 0; i < nodesPerGroup; i++) {
+                Node node = new Node(distributionKey, "host" + distributionKey, group);
+                node.setWorking(true);
+                groupNodes.add(node);
+                distributionKey++;
+            }
+            Group g = new Group(group, groupNodes);
+            groups.put(group, g);
+        }
+        return Map.copyOf(groups);
     }
 
 }
